# MondoBacman-swagger.json
{
  "openapi" : "3.0.0",
  "servers" : [ {
    "description" : "Localhost default server",
    "url" : "http://localhost:5000"
  } ],
  "info" : {
    "description" : "MonDo-BaCMan API for DITAS",
    "version" : "1.0.0",
    "title" : "MonDoBacMan",
    "contact" : {
      "email" : "vpsomak@mail.ntua.gr"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "paths" : {
...
    "/manager/getClusterMap" : {
      "post" : {
        "summary" : "Gets the current cluster map from a MongoDB cluster access point.",
        "operationId" : "getClusterMap",
        "description" : "By passing in the access point to a MongoDB cluster, the manager requests\nthe current cluster map as the mongos instance reads it. This includes\nall the participating configuration and sharding members with their\nreplica set IDs, IPs and ports.\n",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MongoAPConfig"
              }
            }
          },
          "description" : "Configuration of the cluster access point."
        },
        "responses" : {
          "200" : {
            "description" : "The cluster map object.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ClusterMapOutput"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "MongoAPConfig" : {
        "type" : "object",
        "required" : [ "username", "mongo_server", "mongo_port" ],
        "properties" : {
          "mongo_server" : {
            "type" : "string",
            "example" : "localhost",
            "description" : "The IP or domain name of the MongoDB cluster access point."
          },
          "mongo_port" : {
            "type" : "integer",
            "example" : 27017,
            "description" : "The Port of the MongoDB cluster access point."
          },
          "username" : {
            "type" : "string",
            "example" : "John_Doe",
            "description" : "The username of the creator of the containers."
          }
        }
      },
...
      "ClusterMapOutput" : {
        "properties" : {
          "config" : {
            "type" : "string",
            "example" : "rsC/192.168.1.15:27018,192.168.1.15:27019",
            "description" : "The replica set ID, the IPs and the ports of all configuration servers in this cluster."
          },
          "rsID" : {
            "type" : "string",
            "example" : "rsID/192.168.1.15:27020,192.168.1.15:27021",
            "description" : "The replica set ID, the IPs and the ports of all sharding servers in this replica set."
          },
          "rsID2" : {
            "type" : "string",
            "example" : "rsID2/192.168.1.16:27020,192.168.1.16:27021",
            "description" : "The replica set ID, the IPs and the ports of all sharding servers in this replica set."
          },
          "rsIDX" : {
            "type" : "string",
            "example" : "rsIDX/192.168.1.1X:27020,192.168.1.1X:27021",
            "description" : "The replica set ID, the IPs and the ports of all sharding servers in this replica set."
          }
        },
        "type" : "object"
      },
...
    }
  }
}